Move and route augment_paragraphs through augmentor
