+++ nucliadb/src/nucliadb/search/augmentor/__init__.py
@@ -18,4 +18,4 @@
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 #
 from . import fields, paragraphs, resources  # noqa: F401
-from .augmentor import augment  # noqa: F401
+from .augmentor import augment, augment_paragraphs  # noqa: F401
