+++ nucliadb/src/nucliadb/search/search/find_merge.py
@@ -25,7 +25,7 @@ from nidx_protos.nodereader_pb2 import GraphSearchResponse, SearchResponse
 from nucliadb.common.external_index_providers.base import TextBlockMatch
 from nucliadb.common.ids import ParagraphId
 from nucliadb.models.internal.augment import AugmentedParagraph, Paragraph, ParagraphText
-from nucliadb.search.augmentor.paragraphs import augment_paragraphs
+from nucliadb.search.augmentor import augment_paragraphs
 from nucliadb.search.augmentor.resources import augment_resources_deep
 from nucliadb.search.search.cut import cut_page
 from nucliadb.search.search.hydrator import (
