+++ nucliadb/tests/search/unit/test_find_post_index.py
@@ -182,7 +182,7 @@ async def test_find_post_index_search(expected_find_response: dict[str, Any], pr
             side_effect=mock_augment_resources,
         ),
         patch(
-            "nucliadb.search.augmentor.paragraphs.augment_paragraph",
+            "nucliadb.search.augmentor.augmentor.augment_paragraph",
             side_effect=mock_augment_paragraph,
         ),
         patch.object(predict_mock, "rerank", AsyncMock(side_effect=fake_reranking)),
